Reduced the memory usage when allocating partial solvation terms. Should greatly reduce memory usage especially when using CCD.
Added flags so that only certain perturbations can be turned on. Use the
flag "allowXXX false" where XXX is the name of the perturbation.
Available types of perturbations are "SSNE,SSCE,LCA,PF,SHEAR,BACKRUB".

Fixed the genStructDEE protocol so that more than one conformation can
be enumerated.

Partial pairwise matrices are now written to a file instead of sent
through MPI from the slave to the master. Please make sure that the
matrix is written to a temporary directory so these partial matrices
don't fill up normal HD space. I should delete them once the full matrix
is made, but haven't gotten to that yet.

IMPORTANT: Fixed problem with incorrect calculation of entropy scores
when full conformation is enumerated.