Amber96ext/BakerHBond/HBondEnergy: Changed the hydrogen bond potential to use an if statement to keep track of per residue energies instead of using a separate function.
HBondEnergy: Added code to automatically calculate the "neighbor" atom
for non-amino acids.

KSParser: Changed handleMinDEEApplyRot (function that creates PDBs given
a conf file) to properly handle a run that uses an expanded rotamer
library. Also changed the function to properly incorporate reference
energies if they were calculated and used for the design run. 

Metrics: Prints out the boundTightnessPerLevel (measure of how tight the
bound is per A* tree level) if it was calculated.

Molecule: Added a hard coded line to skip handedness check for folate
because it has all atoms an amino acid has, but they do not correspond
to and amino acid peptide. IMPORTANT: This check needs to be fixed to
skip the handedness check for all non-amino acids.

MutUtils: Fixed a bug where a mutation from "DGLY" was not flagged as a
Gly mutation. Changed where the coordinates for the WT rotamer are
stored. From MutUtils to when the mutable residue is initialized. 

PGgurobiAStar: Changed the default preexpanded domain size from 2 to 1.
Fixed the hmeanstatic and hmeandynamic functions so that they properly
calculate the harmonic mean. Save and reset the "flexible" flag for each
residue in the molecule (when EPIC is run the allowed flexible residues
are changed to calculate EPIC bounds during A*, so they must be reset at
the end of the A* search). Compute the fScore of the root node so that
no node has an negative infinity energy (needed by hmeanstatic when
normalizing against the parent node).  Added code that will find the
root node bound and the actual GMEC and then exit (was used to find the
tightness of each f-score technique but is currently commented out
because it is not needed for actual runs). 

Residue: The coordinates for the WT rotamer are now saved when the
mutable residue is initialized.

RotamerSearch: Added flag "savePDBs" that will save low energy pdbs
during the conf enumeration step. Added the ability to pass a max number
of conformations that should be enumerated before the conf enumeration
should return. 

Settings: Added flag "savePDBs" that will save low energy pdbs during
the conf enumeration step.